+++ launch_citra.bat
@@ -4,14 +4,22 @@ setlocal enableextensions
 Color 0A
 cls
 title Portable Citra Launcher - Helper Edition
-set nag=BE SURE TO TURN CAPS LOCK OFF! (never said it was on just make sure)
+set nag=Finally Getting Updates After 4 Years (Helper Update)
 set new_version=OFFLINE_OR_NO_UPDATES
-if exist replacer.bat del replacer.bat
-if exist "%~n0_poc.bat" del "%~n0_poc.bat"
+
+set "name=%~n0"
+set "name=!name:launch_=!"
+set "license=.\doc\!name!_license.txt"
+set "main_launcher=%~n0.bat"
+set "poc_launcher=%~n0_poc.bat"
+set "quick_launcher=quick%~n0.bat"
+
+if exist replacer.bat del replacer.bat >nul:
+if exist !poc_launcher! del !poc_launcher! >nul:
 set "folder=%CD%"
 if "%CD%"=="%~d0\" set "folder=%CD:~0,2%"
 
-call :Alpha-To-Number
+call :AlphaToNumber
 call :SetArch
 call :FolderCheck
 call :Version
@@ -43,7 +51,7 @@ echo e. install text-reader [update if had]
 echo.
 set /p choice="enter a number and press enter to confirm: "
 :: sets errorlevel to 0 (?)
-ver > nul
+ver >nul:
 :: an incorrect call throws an errorlevel of 1
 :: replace all goto Main with (goto) 2>nul (if they are called by the main menu)
 call :%choice%
@@ -64,7 +72,7 @@ call :UpgradeCitra
 
 :2
 :Launch-Citra
-if not exist ".\bin\citra\citra-qt.exe" set "nag=PLEASE INSTALL CITRA FIRST" && (goto) 2>nul
+if not exist ".\bin\citra\citra-qt.exe" set "nag=PLEASE INSTALL CITRA FIRST" & (goto) 2>nul
 title DO NOT CLOSE
 set "path=!PATH!;!folder!\dll\64\;"
 cls
@@ -83,17 +91,17 @@ rmdir /s /q .\data\AppData\Roaming\Citra\
 cls
 taskkill /f /im citra-qt.exe
 rmdir /s /q .\bin\citra\
-del .\extra\citra-setup-windows.exe
+if exist .\extra\citra-setup=windows.exe del .\extra\citra-setup-windows.exe >nul:
 (goto) 2>nul
 
 :5
 :UpdateCheck
-if exist version.txt del version.txt
+if exist version.txt del version.txt >nul:
 cls
 title Portable Citra Launcher - Helper Edition - Checking For Update
 call :HelperDownload "https://raw.githubusercontent.com/MarioMasta64/EverythingPortable/master/version.txt" "version.txt"
 set Counter=0 & for /f "DELIMS=" %%i in ('type version.txt') do (set /a Counter+=1 & set "Line_!Counter!=%%i")
-if exist version.txt del version.txt
+if exist version.txt del version.txt >nul:
 set new_version=%Line_36%
 if "%new_version%"=="OFFLINE" call :ErrorOffline & (goto) 2>nul
 if %current_version% EQU %new_version% call :LatestBuild & (goto) 2>nul
@@ -105,8 +113,8 @@ call :ErrorOffline & (goto) 2>nul
 :6
 :About
 cls
-del .\doc\citra_license.txt
-start launch_citra.bat
+if exist !license! del !license! >nul:
+start %~n0
 exit
 
 :7
@@ -116,7 +124,7 @@ exit
 :DLLDownloaderCheck
 cls & title Portable Citra Launcher - Helper Edition - Download Dll Downloader
 call :HelperDownload "https://raw.githubusercontent.com/MarioMasta64/DLLDownloaderPortable/master/launch_dlldownloader.bat" "launch_dlldownloader.bat.1"
-cls & if exist launch_dlldownloader.bat.1 del launch_dlldownloader.bat & rename launch_dlldownloader.bat.1 launch_dlldownloader.bat
+cls & if exist launch_dlldownloader.bat.1 del launch_dlldownloader.bat >nul: & rename launch_dlldownloader.bat.1 launch_dlldownloader.bat
 cls & start launch_dlldownloader.bat
 (goto) 2>nul
 
@@ -124,7 +132,7 @@ cls & start launch_dlldownloader.bat
 :PortableEverything
 cls & title Portable Citra Launcher - Helper Edition - Download Suite
 call :HelperDownload "https://raw.githubusercontent.com/MarioMasta64/EverythingPortable/master/launch_everything.bat" "launch_everything.bat.1"
-cls & if exist launch_everything.bat.1 del launch_everything.bat & rename launch_everything.bat.1 launch_everything.bat
+cls & if exist launch_everything.bat.1 del launch_everything.bat >nul: & rename launch_everything.bat.1 launch_everything.bat
 cls & start launch_everything.bat
 (goto) 2>nul
 
@@ -132,25 +140,26 @@ cls & start launch_everything.bat
 :Quicklauncher-Check
 cls
 title Portable Citra Launcher - Helper Edition - Quicklauncher Writer
-echo @echo off > quicklaunch_citra.bat
-echo Color 0A >> quicklaunch_citra.bat
-echo cls >> quicklaunch_citra.bat
-echo set "folder=%%CD%%" >> quicklaunch_citra.bat
-echo if "%%CD%%"=="%%~d0\" set "folder=%%CD:~0,2%%" >> quicklaunch_citra.bat
-echo set "UserProfile=%%folder%%\data\" >> quicklaunch_citra.bat
-echo set "AppData=%%folder%%\data\AppData\Roaming\" >> quicklaunch_citra.bat
-echo set "LocalAppData=%%folder%%\data\AppData\Local\" >> quicklaunch_citra.bat
-echo cls >> quicklaunch_citra.bat
-echo start .\bin\citra\citra-qt.exe >> quicklaunch_citra.bat
-echo exit >> quicklaunch_citra.bat
-echo A QUICKLAUNCHER HAS BEEN WRITTEN TO: quicklaunch_citra.bat
-echo ENTER TO CONTINUE & pause>nul:
-(goto) 2>nul
+echo @echo off >!quick_launcher!
+echo Color 0A >>!quick_launcher!
+echo cls >>!quick_launcher!
+echo set "folder=%%CD%%" >>!quick_launcher!
+echo if "%%CD%%"=="%%~d0\" set "folder=%%CD:~0,2%%" >>!quick_launcher!
+echo set "UserProfile=%%folder%%\data" >>!quick_launcher!
+echo set "AppData=%%folder%%\data\AppData\Roaming" >>!quick_launcher!
+echo set "LocalAppData=%%folder%%\data\AppData\Local" >>!quick_launcher!
+echo set "ProgramData=%%folder%%\data\ProgramData" >>!quick_launcher!
+echo cls >>!quick_launcher!
+echo start .\bin\citra\citra-qt.exe >>!quick_launcher!
+echo exit >>!quick_launcher!
+echo A QUICKLAUNCHER HAS BEEN WRITTEN TO:!quick_launcher!
+echo ENTER TO CONTINUE & pause >nul:
+exit
 
 :d
 :UpgradeCitra
 title Portable Citra Launcher - Helper Edition - Citra Update Check
-if exist win!arch!_portable.zip del win!arch!_portable.zip
+if exist citra-setup-windows.exe del citra-setup-windows.exe >nul:
 call :HelperDownload "https://github.com/citra-emu/citra-web/releases/download/1.0/citra-setup-windows.exe" "citra-setup-windows.exe"
 :MoveCitra
 move citra-setup-windows.exe .\extra\citra-setup-windows.exe
@@ -158,16 +167,17 @@ move citra-setup-windows.exe .\extra\citra-setup-windows.exe
 cls
 echo continue with installation,
 echo when prompted for location enter,
+echo do not choose canary
 echo "!folder!\bin\citra"
-pause
+echo ENTER TO CONTINUE & pause >nul:
 .\extra\citra-setup-windows.exe
 :CleanupCitra
 cls
 for /d %%i in (".\bin\citra\*") do if /i not "%%i"==".\bin\citra\nightly-mingw" rmdir /s /q "%%i"
-for /R %%i in (.\bin\citra\*) do del "%%i"
+for /R %%i in (.\bin\citra\*) do del "%%i" >nul:
 xcopy .\bin\citra\nightly-mingw\* .\bin\citra\ /e /i /y
-rmdir /s /q .\bin\citra\nightly-mingw\
-rmdir /s /q "!AppData!\Microsoft\Windows\Start Menu\Programs\Citra\"
+if exist .\bin\citra\nightly-mingw\ rmdir /s /q .\bin\citra\nightly-mingw\
+if exist "!AppData!\Microsoft\Windows\Start Menu\Programs\Citra\" rmdir /s /q "!AppData!\Microsoft\Windows\Start Menu\Programs\Citra\"
 (goto) 2>nul
 
 :e
@@ -183,45 +193,62 @@ REM STUFF THAT IS ALMOST IDENTICAL BETWEEN STUFF
 
 :FolderCheck
 cls
-set "UserProfile=!folder!\data\"
-set "AppData=!folder!\data\AppData\Roaming\"
-set "LocalAppData=!folder!\data\AppData\Local\"
+set "UserProfile=!folder!\data"
+set "AppData=!folder!\data\AppData\Roaming"
+set "LocalAppData=!folder!\data\AppData\Local"
+set "ProgramData=!folder!\data\ProgramData"
 if not exist .\bin\ mkdir .\bin\
 if not exist .\data\ mkdir .\data\
 if not exist .\doc\ mkdir .\doc\
 if not exist .\extra\ mkdir .\extra\
 if not exist .\helpers\ mkdir .\helpers\
+if not exist .\ini\ mkdir .\ini\
 if not exist .\note\ mkdir .\note\
 if not exist .\data\AppData\Local\ mkdir .\data\AppData\Local\
 if not exist .\data\AppData\Roaming\ mkdir .\data\AppData\Roaming\
+if not exist .\data\ProgramData\ mkdir .\data\ProgramData\
+if not exist ".\data\3D Objects\" mkdir ".\data\3D Objects\"
+if not exist ".\data\Contacts\" mkdir ".\data\Contacts\"
+if not exist ".\data\Desktop\" mkdir ".\data\Desktop\"
+if not exist ".\data\Documents\" mkdir ".\data\Documents\"
+if not exist ".\data\Downloads\" mkdir ".\data\Downloads\"
+if not exist ".\data\Favorites\" mkdir ".\data\Favorites\"
+if not exist ".\data\Links\" mkdir ".\data\Links\"
+if not exist ".\data\Music\" mkdir ".\data\Music\"
+if not exist ".\data\OneDrive\" mkdir ".\data\OneDrive\"
+if not exist ".\data\Pictures\" mkdir ".\data\Pictures\"
+if not exist ".\data\Saved Games\" mkdir ".\data\Saved Games\"
+if not exist ".\data\Searches\" mkdir ".\data\Searches\"
+if not exist ".\data\Videos\" mkdir ".\data\Videos\"
 if not exist ".\bin\citra\citra-qt.exe" set nag=CITRA IS NOT INSTALLED CHOOSE "D"
 (goto) 2>nul
 
 :Version
 cls
-echo 2 > .\doc\version.txt
+echo 3 > .\doc\version.txt
 set /p current_version=<.\doc\version.txt
-if exist .\doc\version.txt del .\doc\version.txt
-:: REPLACE ALL exit /b that dont need an error code (a value after it) with "exit"
+if exist .\doc\version.txt del .\doc\version.txt >nul:
 (goto) 2>nul
 
 :Credits
 cls
-if exist .\doc\citra_license.txt (goto) 2>nul
-echo ================================================== > .\doc\citra_license.txt
-echo =              Script by MarioMasta64            = >> .\doc\citra_license.txt
-echo =           Script Version: v%current_version%- release         = >> .\doc\citra_license.txt
-echo ================================================== >> .\doc\citra_license.txt
-echo =You may Modify this WITH consent of the original= >> .\doc\citra_license.txt
-echo = creator, as long as you include a copy of this = >> .\doc\citra_license.txt
-echo =      as you include a copy of the License      = >> .\doc\citra_license.txt
-echo ================================================== >> .\doc\citra_license.txt
-echo =    You may also modify this script without     = >> .\doc\citra_license.txt
-echo =         consent for PERSONAL USE ONLY          = >> .\doc\citra_license.txt
-echo ================================================== >> .\doc\citra_license.txt
+if exist !license! (goto) 2>nul
+echo ================================================== > !license!
+echo =              Script by MarioMasta64            = >> !license!
+set "extra_space="
+if %current_version% LSS 10 set "extra_space= "
+echo =           Script Version: v%current_version%- release        %extra_space%= >> !license!
+echo ================================================== >> !license!
+echo =You may Modify this WITH consent of the original= >> !license!
+echo = creator, as long as you include a copy of this = >> !license!
+echo =      as you include a copy of the License      = >> !license!
+echo ================================================== >> !license!
+echo =    You may also modify this script without     = >> !license!
+echo =         consent for PERSONAL USE ONLY          = >> !license!
+echo ================================================== >> !license!
 cls
 title Portable Citra Launcher - Helper Edition - About
-for /f "DELIMS=" %%i in (.\doc\citra_license.txt) do (echo %%i)
+for /f "DELIMS=" %%i in (!license!) do (echo %%i)
 pause
 call :PingInstall
 (goto) 2>nul
@@ -239,7 +266,7 @@ if not exist launch_helpers.bat call :DownloadHelpers
 (goto) 2>nul
 :DownloadHelpers
 if not exist .\helpers\download.vbs call :CreateDownloadVBS
-cscript .\helpers\download.vbs https://raw.githubusercontent.com/MarioMasta64/EverythingPortable/master/launch_helpers.bat launch_helpers.bat > nul
+cscript .\helpers\download.vbs https://raw.githubusercontent.com/MarioMasta64/EverythingPortable/master/launch_helpers.bat launch_helpers.bat >nul:
 (goto) 2>nul
 :CreateDownloadVBS
 echo Dim Arg, download, file > .\helpers\download.vbs
@@ -327,9 +354,11 @@ echo %sha1%
 set program=%~n0
 echo %program:~7%
 echo "https://mariomasta64.me/install/new_install.php?program=%program:~7%^&serial=%sha1%"
+if exist new_install.php del new_install.php >nul:
+if exist serial.txt del serial.txt >nul:
 REM call :HelperDownload "https://mariomasta64.me/install/new_install.php?program=%program:~7%^&serial=%sha1%" "new_install.php"
-del new_install.php*
-del serial.txt
+if exist new_install.php del new_install.php >nul:
+if exist serial.txt del serial.txt >nul:
 (goto) 2>nul
 
 :UpdateWget
@@ -343,8 +372,8 @@ title Portable Citra Launcher - Helper Edition - Latest Build :D
 echo you are using the latest version!!
 echo Current Version: v%current_version%
 echo New Version: v%new_version%
-echo ENTER TO CONTINUE & pause>nul:
-start launch_citra.bat
+echo ENTER TO CONTINUE & pause >nul:
+start %~n0
 exit
 
 :NewUpdate
@@ -364,8 +393,8 @@ goto NewUpdate
 
 :UpdateNow
 cls & title Portable Citra Launcher - Helper Edition - Updating Launcher
-call :HelperDownload "https://raw.githubusercontent.com/MarioMasta64/EverythingPortable/master/launch_citra.bat" "launch_citra.bat.1"
-cls & if exist launch_citra.bat.1 goto ReplacerCreate
+call :HelperDownload "https://raw.githubusercontent.com/MarioMasta64/EverythingPortable/master/!main_launcher!" "!main_launcher!.1"
+cls & if exist %~n0.1 goto ReplacerCreate
 cls & call :ErrorOffline
 (goto) 2>nul
 
@@ -373,11 +402,11 @@ cls & call :ErrorOffline
 cls
 echo @echo off > replacer.bat
 echo Color 0A >> replacer.bat
-echo del launch_citra.bat >> replacer.bat
-echo rename launch_citra.bat.1 launch_citra.bat >> replacer.bat
-echo start launch_citra.bat >> replacer.bat
+echo del %~n0 >> replacer.bat
+echo rename %~n0.1 %~n0 >> replacer.bat
+echo start %~n0 >> replacer.bat
 :: launcher exits, deletes itself, and then exits again. yes. its magic.
-echo (goto) 2^>nul ^& del "%%~f0" ^& exit >> replacer.bat
+echo (goto) 2^ >nul: ^& del "%%~f0" ^& exit >> replacer.bat
 call :HelperHide "replacer.bat"
 exit
 
@@ -388,8 +417,8 @@ echo YOURE USING A TEST BUILD MEANING YOURE EITHER
 echo CLOSE TO ME OR YOURE SOME SORT OF PIRATE
 echo Current Version: v%current_version%
 echo New Version: v%new_version%
-echo ENTER TO CONTINUE & pause>nul:
-start launch_citra.bat
+echo ENTER TO CONTINUE & pause >nul:
+start %~n0
 exit
 
 :ErrorOffline
@@ -446,7 +475,7 @@ echo (C) Copyright Microsoft Corporation. All rights reserved
 echo.
 echo nice job finding me. have fun with my little cmd prompt.
 echo upon error (more likely than not) i will return to the menu.
-echo type "(goto) 2^>nul" or make me error to return.
+echo type "(goto) 2^ >nul:" or make me error to return.
 echo.
 :CmdLoop
 set /p "cmd=%cd%>"
@@ -461,13 +490,13 @@ echo cls >> relaunch.bat
 echo Color 0A >> relaunch.bat
 echo start %~f0 >> relaunch.bat
 :: launcher exits, deletes itself, and then exits again. yes. its magic.
-echo (goto) 2^>nul ^& del "%%~f0" ^& exit >> relaunch.bat
+echo (goto) 2^ >nul: ^& del "%%~f0" ^& exit >> relaunch.bat
 call :HelperHide "relaunch.bat"
 exit
 
 REM extra stuff ill remove one day
 
-replace launch_citra.bat with %~f0
+replace %~n0 with %~f0
 because i use the call command. you can edit the file add a label and goto the label by typing it in the menu without even having to close the program cause youre worried about it glitching (put your code on the bottom)
 add raw before raw/master in everything
 maybe add option to open mod folder?
